pyproject.toml
[project]
name = "veille-parlementaire-lidl"
version = "0.1.0"
description = "Veille parlementaire et institutionnelle automatisée — compte Lidl France / grande distribution alimentaire (Sideline Conseil)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "feedparser>=6.0",
  "lxml>=5.2",
  "beautifulsoup4>=4.12",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "tenacity>=9.0",
  "jinja2>=3.1",
  "python-dateutil>=2.9",
  "unidecode>=1.3",
  "rich>=13.7",
  # R18 (2026-04-22) : curl_cffi fait des requêtes avec la signature TLS
  # de Chrome — passe les protections Cloudflare JS-challenge (ministères
  # education/interieur/economie/justice + info.gouv, qui bloquent httpx
  # standard en 403). Fallback silencieux dans _common si import KO.
  "curl_cffi>=0.7",
  # R35-B (2026-04-24) : extraction texte des PDFs des comptes rendus de
  # commissions AN. pypdf est un fallback léger (pas de deps natives) utile
  # pour parser /dyn/17/comptes-rendus/{slug}/l17..._compte-rendu.pdf sans
  # avoir à installer libpoppler. Rendu best-effort : si l'extraction
  # échoue, le matcher retombe sur le titre seul.
  "pypdf>=4.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.5"]

[project.scripts]
veille = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 110
target-version = "py311"
